add default ffi namespce
