+++ python/dgl/_ffi/capi.py
@@ -1,0 +1,6 @@
+"""Init all C APIs in the default namespace."""
+from .function import _init_api
+
+__all__ = _init_api("dgl.capi", __name__)
+
+print(__all__)
