+++ python/dgl/_ffi/function.py
@@ -295,16 +295,18 @@ def _init_api(namespace, target_module_name=None):
     target_module_name : str
        The target module name if different from namespace
     """
+    print(target_module_name)
     target_module_name = target_module_name if target_module_name else namespace
     if namespace.startswith("dgl."):
-        _init_api_prefix(target_module_name, namespace[4:])
+        return _init_api_prefix(target_module_name, namespace[4:])
     else:
-        _init_api_prefix(target_module_name, namespace)
+        return _init_api_prefix(target_module_name, namespace)
 
 
 def _init_api_prefix(module_name, prefix):
     module = sys.modules[module_name]
-
+    name_list = []
+    
     for name in list_global_func_names():
         if name.startswith("_") and not name.startswith("_deprecate"):
             # internal APIs are ignored
@@ -324,8 +326,9 @@ def _init_api_prefix(module_name, prefix):
         ff.__name__ = fname
         ff.__doc__ = "DGL PackedFunc %s. " % fname
         setattr(target_module, ff.__name__, ff)
-
-
+        name_list.append(fname)
+    return name_list
+    
 def _init_internal_api():
     for name in list_global_func_names():
         if not name.startswith("_") or name.startswith("_deprecate"):
