+++ python/dgl/transforms/__init__.py
@@ -1,3 +1,4 @@
 """Transform for structures and features"""
 from .functional import *
 from .module import *
+from .to_block import *
